+++ src/main/java/com/butchercraft/machine/grinder/GrinderBlockEntity.java
@@ -1,7 +1,8 @@
 package com.butchercraft.machine.grinder;
 
-import com.butchercraft.registration.ModBlockEntityTypes;
 import com.butchercraft.product.integration.DevelopmentProductItemMappings;
+import com.butchercraft.registration.ModBlockEntityTypes;
+import com.butchercraft.workstation.WorkstationExecutionStrategy;
 import com.butchercraft.workstation.WorkstationOperationResolver;
 import com.butchercraft.workstation.block.AbstractProcessingWorkstationBlockEntity;
 import net.minecraft.core.BlockPos;
@@ -21,7 +22,8 @@ public final class GrinderBlockEntity extends AbstractProcessingWorkstationBlock
                 blockState,
                 GrinderWorkstation.capability(),
                 new WorkstationOperationResolver(),
-                DevelopmentProductItemMappings.fixtureMapping()
+                DevelopmentProductItemMappings.fixtureMapping(),
+                WorkstationExecutionStrategy.transformation()
         );
     }